move Annex B tests to preserve-use-strict-cjs.test.ts
runtime-transpiler.test.ts has a pre-existing release-mode failure in the
`// @bun` handlebars tests that is unrelated to this change (a transpiler
cache hit on the main module skips the has_loaded bookkeeping that drives
the unknown-extension loader fallback). The sloppy/strict-mode CJS file is
the more natural home for these tests anyway.
