+++ test/bundler/transpiler/runtime-transpiler.test.ts
@@ -195,127 +195,6 @@ describe("with statement", () => {
   });
 });
 
-describe("Annex B.3.3 block-level function hoisting in sloppy mode", () => {
-  const fixture = `
-{ function z() { return 9 } }
-console.log(z());
-if (true) function u() { return 1 }
-console.log(typeof u);
-if (true) { function w() { return 2 } }
-console.log(typeof w);
-function outer() {
-  { function inner() { return "in" } }
-  return inner();
-}
-console.log(outer());
-`;
-
-  test.concurrent("hoists block-level functions to the enclosing var scope in CJS", async () => {
-    using dir = tempDir("annex-b-block-fn-cjs", {
-      "z.cjs": fixture,
-    });
-    await using proc = Bun.spawn({
-      cmd: [bunExe(), "z.cjs"],
-      env: bunEnv,
-      cwd: String(dir),
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-    expect(stderr).toBe("");
-    expect(stdout).toBe("9\nfunction\nfunction\nin\n");
-    expect(exitCode).toBe(0);
-  });
-
-  test.concurrent("keeps block-level functions block-scoped in strict mode (ESM)", async () => {
-    using dir = tempDir("annex-b-block-fn-esm", {
-      "z.mjs": `{ function z() { return 9 } }\nconsole.log(typeof z);\n`,
-    });
-    await using proc = Bun.spawn({
-      cmd: [bunExe(), "z.mjs"],
-      env: bunEnv,
-      cwd: String(dir),
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-    expect(stderr).toBe("");
-    expect(stdout).toBe("undefined\n");
-    expect(exitCode).toBe(0);
-  });
-
-  test.concurrent("does not hoist past a conflicting lexical binding", async () => {
-    using dir = tempDir("annex-b-block-fn-conflict", {
-      "z.cjs": `let z = 1;\n{ function z() {} }\nconsole.log(typeof z);\n`,
-    });
-    await using proc = Bun.spawn({
-      cmd: [bunExe(), "z.cjs"],
-      env: bunEnv,
-      cwd: String(dir),
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-    expect(stderr).toBe("");
-    expect(stdout).toBe("number\n");
-    expect(exitCode).toBe(0);
-  });
-
-  // https://github.com/oven-sh/bun/issues/23633
-  test.concurrent("assigns the hoisted var when a block-level function shadows an outer one", async () => {
-    using dir = tempDir("annex-b-block-fn-shadow", {
-      "foo.cjs": `
-function foo() { console.log('foo') }
-foo();
-{
-  function foo() { console.log('bar') }
-  foo();
-}
-foo();
-`,
-    });
-    await using proc = Bun.spawn({
-      cmd: [bunExe(), "foo.cjs"],
-      env: bunEnv,
-      cwd: String(dir),
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-    const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-    expect(stderr).toBe("");
-    expect(stdout).toBe("foo\nbar\nbar\n");
-    expect(exitCode).toBe(0);
-  });
-
-  // https://github.com/oven-sh/bun/issues/25737
-  test.concurrent("hoists a labeled function declaration to the enclosing var scope", async () => {
-    using dir = tempDir("annex-b-labeled-fn", {
-      "package.json": `{"type":"commonjs"}`,
-      "top.js": `foo:\n    function bar() { return "bar"; }\nconsole.log(bar());\n`,
-      "nested.js": `
-function lex() {
-  loop: function _lex() { return "ok"; }
-  return _lex();
-}
-console.log(lex());
-`,
-    });
-    for (const file of ["top.js", "nested.js"]) {
-      await using proc = Bun.spawn({
-        cmd: [bunExe(), file],
-        env: bunEnv,
-        cwd: String(dir),
-        stdout: "pipe",
-        stderr: "pipe",
-      });
-      const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-      expect(stderr).toBe("");
-      expect(stdout).toBe(file === "top.js" ? "bar\n" : "ok\n");
-      expect(exitCode).toBe(0);
-    }
-  });
-});
-
 test("math.pow", () => {
   function foo1(foo) {
     return 10 ** (foo / 20);
